Moved PyHoc namespace following code review by @ramcdougal.
